Refactor experimental features to internal package
